requirements.txt
aoi-queueing-tool==1.2.4
aoi-simulator==0.3.2
autograd==1.4
autograd-gamma==0.5.0
cycler==0.11.0
docutils==0.17.1
fonttools==4.33.1
future==0.18.2
kiwisolver==1.4.2
matplotlib==3.5.1
mplcursors==0.5.1
networkx==2.8
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pkg_resources==0.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
reliability==0.8.3
scipy==1.8.0
six==1.16.0
